fix decorating Symfony's ProblemNormalizer
The previous definition was overriding the tag which removed the priority set on it by the decorated service.
The consequence is that this normalizer had a priority of `0` and could change the behavior of an application. For
instance when a project uses API Platform: their own error normalizer has a lower priority from the Symfony's one but
was never used because of this change in the prioirities of normalizers.