fix: use isolated context in tests
Since the `Context` class uses the global scope we must isolate the contexts used in each test so that there
is no conflict.

By forking the main context we will work in an identical copy.
When a test finishes we can destroy this fork and switch back to the main context.