chore: make data collector independent from OpenTelemetry SDK
We got errors because the `data` attribute is serialized, using PHP
`serialize` function, by the parent class.
Passing objects to this array is risky since we don't control what is in
there.
Instead, we know only use it to build the view expected by the twig
template.

The issue happened on a project using Guzzle as HTTP client. It was used
as transport for the span.
The transport is part of each "span object", in the case of Guzzle: they
use closures to handle requests.
Because PHP can't serialize a closure and was trying to serialize the
span and its exporter, it was failing.