feat: include previous exception information in normalized stacktrace
PHP `\Throwable::getTraceAsString()` implementation only return the
stack trace of the current exception object.
It doesn't include any information from the previous expcetion(s).

But `\Throwable::__toString()` does, it will show each exception type,
message and stacktrace. From the oldest to the newest.

This is valuable information when looking at the logs.
Let's imagine we receive a request to sign in a new user, this request
is validated and the validation fails because the password is too short.
And that we wrap this error, later in the process, in a more generic
error with the message "invalid credentials" error.
Currently, all what the logs are gonna let us know is that the
credentials were invalid.
With this proposal, we will also be able to know that the password was
too short.