fix: use doctrine version aware drivers
When the drivers don't implement this interface, Doctrine is not able to
know that it can create the platform for a specific version.
It seems to fallback to using the default platform.
For instance, with MySQL, the "json" type is not handled in all
versions, the generic driver doesn't handle it.
Not using a driver aware of the version generates error when using CLI
commands for doctrine: "Unknown database type json requested,
Doctrine\DBAL\Platforms\MySQLPlatform may not support it."