chore: add span to response user data
The handling of the `on_response` callback is currently blocking instead
of being asynchronous.
Overall we believe the callbacks where not the right answer and that
making the span available through the response is a better way forward.
Each clients will be free to access it trigger behavior on
request/response without having to depend on the `TracingHttpClient`
implementation.
Removing this coupling will make our client easier to maintain by
removing an undesirable responsibility.